Don't allow question marks or periods in Zendesk subdomain
For security reasons, question marks and periods should not be present
in the Zendesk subdomain. If a subdomain value such as
"bishopfox.com?name=value" is passed, with the current redirect logic
we'd send the user to "https://bishopfox.com?name=value.zendesk.com",
which could be an attacker-controlled domain.